+++ vod/media_set.h
@@ -182,7 +182,7 @@ typedef struct {
 	uint32_t tracks_mask[MEDIA_TYPE_COUNT];
 	sequence_tracks_mask_t* sequence_tracks_mask;
 	sequence_tracks_mask_t* sequence_tracks_mask_end;
-	uint8_t* langs_mask;			// [LANG_MASK_SIZE]
+	uint64_t* langs_mask;			// [LANG_MASK_SIZE]
 	uint32_t version;
 	uint32_t width;
 	uint32_t height;
